Migrate parts of `MessageUtils` to extension functions
