+++ jvm/middleware/src/test/kotlin/io/spine/logging/jvm/backend/MessageUtilsSpec.kt
@@ -26,25 +26,11 @@
 
 package io.spine.logging.jvm.backend
 
-import io.kotest.matchers.booleans.shouldBeFalse
-import io.kotest.matchers.booleans.shouldBeTrue
 import io.kotest.matchers.shouldBe
-import io.kotest.matchers.string.shouldBeEmpty
 import io.kotest.matchers.string.shouldContain
-import io.kotest.matchers.string.shouldNotContain
 import io.kotest.matchers.types.shouldBeSameInstanceAs
 import io.spine.logging.backend.given.BadToString
-import io.spine.logging.jvm.JvmLogSite
-import io.spine.logging.jvm.backend.FormatOptions.Companion.FLAG_SHOW_ALT_FORM
-import io.spine.logging.jvm.backend.FormatOptions.Companion.FLAG_SHOW_LEADING_ZEROS
-import io.spine.logging.jvm.backend.FormatOptions.Companion.FLAG_UPPER_CASE
-import io.spine.logging.jvm.backend.FormatOptions.Companion.UNSET
-import io.spine.logging.jvm.backend.MessageUtils.appendHex
-import io.spine.logging.jvm.backend.MessageUtils.appendLogSite
-import io.spine.logging.jvm.backend.MessageUtils.safeFormatTo
 import io.spine.logging.jvm.backend.MessageUtils.safeToString
-import io.spine.logging.jvm.given.FakeLogSite
-import java.util.*
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
@@ -57,11 +43,6 @@ import org.junit.jupiter.api.Test
  */
 @DisplayName("`MessageUtils` should")
 internal class MessageUtilsSpec {
-
-    companion object {
-        private val NO_OPTIONS = FormatOptions.getDefault()
-        private val UPPER_CASE = FormatOptions.of(FLAG_UPPER_CASE, UNSET, UNSET)
-    }
 
     @Nested
     inner class
@@ -94,71 +75,4 @@ internal class MessageUtilsSpec {
             safeToString(any) shouldContain "java.lang.IllegalArgumentException: Badness"
         }
     }
-
-    @Nested
-    inner class
-    `safely format the given 'Formattable'` {
-
-        @Test
-        fun `which doesn't throw`() {
-            val out = StringBuilder()
-            val arg = Formattable { formatter, flags, width, precision ->
-                formatter.format(null as Locale?, "[f=%d,w=%d,p=%d]", flags, width, precision)
-            }
-
-            // FormattableFlags.LEFT_JUSTIFY == 1 << 0 = 1
-            // FormattableFlags.UPPERCASE == 1 << 1 = 2
-            // FormattableFlags.ALTERNATE == 1 << 2 = 4
-            safeFormatTo(arg, out, FormatOptions.of(FLAG_UPPER_CASE, 4, 2))
-            "$out" shouldBe "[f=2,w=4,p=2]"
-
-            // Not all flags are passed into the callback.
-            val options = FormatOptions.of(FLAG_SHOW_LEADING_ZEROS + FLAG_SHOW_ALT_FORM, 1, 0)
-            out.setLength(0)
-            safeFormatTo(arg, out, options)
-            "$out" shouldBe "[f=4,w=1,p=0]"
-        }
-
-        @Test
-        fun `which throws an exception`() {
-            val badFormattable = Formattable { formatter, _, _, _ ->
-                formatter.format(null as Locale?, "DISCARDED")
-                throw IllegalArgumentException("Badness")
-            }
-            val out = StringBuilder()
-            safeFormatTo(badFormattable, out, FormatOptions.getDefault())
-            "$out" shouldContain "java.lang.IllegalArgumentException: Badness"
-            "$out" shouldNotContain "DISCARDED"
-        }
-    }
-
-    @Test
-    fun `append log site`() {
-        val out = StringBuilder()
-        val logSite = FakeLogSite("<class>", "<method>", 32, "Ignored.java")
-
-        appendLogSite(logSite, out).shouldBeTrue()
-        "$out" shouldBe "<class>.<method>:32"
-
-        out.setLength(0)
-        appendLogSite(JvmLogSite.INVALID, out).shouldBeFalse()
-        "$out".shouldBeEmpty()
-    }
-
-    @Test
-    fun `append HEX values`() {
-        formatHex(0xFDB97531, NO_OPTIONS) shouldBe "fdb97531"
-        formatHex(0xFDB97531, UPPER_CASE) shouldBe "FDB97531"
-        formatHex(0x0123456789ABCDEFL, NO_OPTIONS) shouldBe "123456789abcdef"
-        formatHex(0x1CDCBA9776543210L, UPPER_CASE) shouldBe "1CDCBA9776543210"
-        formatHex(0, UPPER_CASE) shouldBe "0"
-        formatHex((-1).toByte(), UPPER_CASE) shouldBe "FF"
-        formatHex((-1).toShort(), UPPER_CASE) shouldBe "FFFF"
-    }
-}
-
-private fun formatHex(n: Number, options: FormatOptions): String {
-    val out = StringBuilder()
-    appendHex(out, n, options)
-    return "$out"
 }
